+++ lib/omnibus/fetchers/path_fetcher.rb
@@ -37,10 +37,10 @@ module Omnibus
         # Robocopy's return code is 1 if it succesfully copies over the
         # files and 0 if the files are already existing at the destination
         sync_cmd = "robocopy #{@source[:path]}\\ #{@project_dir}\\ /MIR /S"
-        quiet_shellout!(sync_cmd, returns: [0, 1])
+        shellout!(sync_cmd, returns: [0, 1])
       else
         sync_cmd = "rsync --delete -a #{@source[:path]}/ #{@project_dir}/"
-        quiet_shellout!(sync_cmd)
+        shellout!(sync_cmd)
       end
     end