+++ spec/unit/install_path_cache_spec.rb
@@ -105,7 +105,7 @@ module Omnibus
           .and_return(false)
         expect(FileUtils).to receive(:mkdir_p)
           .with(File.dirname(ipc.cache_path))
-        expect(ipc).to receive(:quiet_shellout!)
+        expect(ipc).to receive(:shellout!)
           .with("git --git-dir=#{cache_path} init -q")
         ipc.create_cache_path
       end
@@ -117,7 +117,7 @@ module Omnibus
         allow(File).to receive(:directory?)
           .with(File.dirname(ipc.cache_path))
           .and_return(true)
-        expect(ipc).to_not receive(:quiet_shellout!)
+        expect(ipc).to_not receive(:shellout!)
           .with("git --git-dir=#{cache_path} init -q")
         ipc.create_cache_path
       end
@@ -125,7 +125,7 @@ module Omnibus
 
     describe '#incremental' do
       before(:each) do
-        allow(ipc).to receive(:quiet_shellout!)
+        allow(ipc).to receive(:shellout!)
         allow(ipc).to receive(:create_cache_path)
       end
 
@@ -135,19 +135,19 @@ module Omnibus
       end
 
       it 'adds all the changes to git' do
-        expect(ipc).to receive(:quiet_shellout!)
+        expect(ipc).to receive(:shellout!)
           .with("git --git-dir=#{cache_path} --work-tree=#{install_path} add -A -f")
         ipc.incremental
       end
 
       it 'commits the backup for the software' do
-        expect(ipc).to receive(:quiet_shellout!)
+        expect(ipc).to receive(:shellout!)
           .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} commit -q -m "Backup of #{ipc.tag}"))
         ipc.incremental
       end
 
       it 'tags the software backup' do
-        expect(ipc).to receive(:quiet_shellout!)
+        expect(ipc).to receive(:shellout!)
           .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} tag -f "#{ipc.tag}"))
         ipc.incremental
       end
@@ -164,10 +164,10 @@ module Omnibus
       end
 
       before(:each) do
-        allow(ipc).to receive(:quiet_shellout)
+        allow(ipc).to receive(:shellout)
           .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} tag -l "#{ipc.tag}"))
           .and_return(tag_cmd)
-        allow(ipc).to receive(:quiet_shellout!)
+        allow(ipc).to receive(:shellout!)
           .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} checkout -f "#{ipc.tag}"))
         allow(ipc).to receive(:create_cache_path)
       end
@@ -178,10 +178,10 @@ module Omnibus
       end
 
       it 'checks for a tag with the software and version, and if it finds it, checks it out' do
-        expect(ipc).to receive(:quiet_shellout)
+        expect(ipc).to receive(:shellout)
           .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} tag -l "#{ipc.tag}"))
           .and_return(tag_cmd)
-        expect(ipc).to receive(:quiet_shellout!)
+        expect(ipc).to receive(:shellout!)
           .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} checkout -f "#{ipc.tag}"))
         ipc.restore
       end
@@ -190,10 +190,10 @@ module Omnibus
         let(:git_tag_output) { "\n" }
 
         it 'does nothing' do
-          expect(ipc).to receive(:quiet_shellout)
+          expect(ipc).to receive(:shellout)
             .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} tag -l "#{ipc.tag}"))
             .and_return(tag_cmd)
-          expect(ipc).to_not receive(:quiet_shellout!)
+          expect(ipc).to_not receive(:shellout!)
             .with(%Q(git --git-dir=#{cache_path} --work-tree=#{install_path} checkout -f "#{ipc.tag}"))
           ipc.restore
         end